services/miner-node/requirements.txt
# Tensora Miner Node - Python Dependencies

# AI Engines
onnxruntime>=1.16.0
vllm>=0.2.7
torch>=2.0.0

# Blockchain
web3>=6.11.0
eth-account>=0.9.0

# IPFS
ipfshttpclient>=0.8.0a2
aiohttp>=3.9.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0
requests>=2.31.0

# Logging
python-json-logger>=2.0.7

# Database (for job queue)
aiosqlite>=0.19.0

